refactor: extract useDOMFinder hook
